Move p2p host validation to CLI validation
